Template and style fixes.
Added conditional display of 'title' and 'body' in posts. This allows modular use to add, for exmaple, a block of text after an image or two images in a row.
Added escaping to 'body' in posts allow html to render. Allows links, italics, etc.
Added exmaples of links and italics in 'body' of posts.
Added styles for links and adjusted timing of transitions.